Switch to seconds and doubles

+++ hal/src/main/native/cpp/jni/DIOJNI.cpp
@@ -133,11 +133,11 @@ Java_edu_wpi_first_hal_DIOJNI_getDIODirection
 /*
  * Class:     edu_wpi_first_hal_DIOJNI
  * Method:    pulse
- * Signature: (II)V
+ * Signature: (ID)V
  */
 JNIEXPORT void JNICALL
 Java_edu_wpi_first_hal_DIOJNI_pulse
-  (JNIEnv* env, jclass, jint id, jint value)
+  (JNIEnv* env, jclass, jint id, jdouble value)
 {
   int32_t status = 0;
   HAL_Pulse((HAL_DigitalHandle)id, value, &status);
@@ -147,11 +147,11 @@ Java_edu_wpi_first_hal_DIOJNI_pulse
 /*
  * Class:     edu_wpi_first_hal_DIOJNI
  * Method:    pulseMultiple
- * Signature: (JI)V
+ * Signature: (JD)V
  */
 JNIEXPORT void JNICALL
 Java_edu_wpi_first_hal_DIOJNI_pulseMultiple
-  (JNIEnv* env, jclass, jlong channelMask, jint value)
+  (JNIEnv* env, jclass, jlong channelMask, jdouble value)
 {
   int32_t status = 0;
   HAL_PulseMultiple(static_cast<uint32_t>(channelMask), value, &status);
